requirements.txt
# Core dependencies
mlflow>=2.0.0
feast>=0.34.0
openlineage-python>=1.0.0

# Database
psycopg2-binary>=2.9.0
redis>=4.0.0

# Web framework
fastapi>=0.100.0
uvicorn>=0.20.0

# Data processing
pandas>=1.5.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Storage
boto3>=1.26.0
minio>=7.0.0

# Utilities
python-dotenv>=1.0.0
requests>=2.28.0
pydantic>=2.0.0

# Development
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.0.0
